backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
beautifulsoup4>=4.12.3
httpx>=0.27.2
python-jose[cryptography]>=3.3.0
redis>=5.2.0
sqlalchemy>=2.0.36
alembic>=1.14.0
prometheus-client>=0.21.0
sentry-sdk[fastapi]>=2.19.0
langchain-core>=0.3.0
pytest>=8.3.4
